Calculate abundances

These functions calculate the abundances of endmembers in the given data using either non-negative least squares (NNLS) or barycentric coordinates. Barycentric coordinates require the intrinsic simplex space, i.e. endmembers must be a (N+1) x N matrix. If you used PCA, consider providing data in PCA scores space. NNLS is more robust and can be applied in the original space. However, the values are not guaranteed to sum to 1. Use the normalize argument to normalize the abundances.

nnls(endmembers, data)

bary(endmembers, data)

abundances(endmembers, data, method = c("nnls", "bary"), normalize = FALSE)

Arguments

endmembers

A matrix where each row is an endmember.

data

A matrix where each row is a spectrum.

method

A character string specifying the method to use for abundance calculation. Options are "nnls" (default) for non-negative least squares and "bary" for barycentric coordinates.

normalize

A logical value indicating whether to normalize the abundances so that they sum to 1.

Value

A matrix where each row is the abundance percentages of the endmembers for the corresponding spectrum.

Examples

data("demo_data")

# Reduce data dimensionality with PCA
pca <- prcomp(demo_data)
x <- pca$x[,1:2]

# Perform N-FINDR in reduced space
nf <- nfindr(x, p = 3)

# Get endmembers both in reduced and original space
ems <- endmembers(nf, demo_data)
ems_pca <- endmembers(nf, x)

# Calculate abundances using barycentric coordinates
# it works only in the reduced space
ab_bary <- abundances(ems_pca, x, method = "bary")

# Calculate abundances using NNLS
# it works in both spaces but it is better to be applied in the original space
ab_nnls <- abundances(ems, demo_data, method = "nnls", normalize = TRUE)

# Alternatively, one can use the specific function directly
nnls(ems, demo_data) # not normalized
#>               [,1]         [,2]         [,3]
#>  [1,] 0.000000e+00 5.185886e-01 3.009660e-01
#>  [2,] 4.896798e-01 2.371365e-01 3.446999e-01
#>  [3,] 1.000000e+00 7.894829e-18 1.457464e-17
#>  [4,] 4.178669e-01 3.988920e-01 2.677794e-01
#>  [5,] 3.829922e-01 1.337541e-01 2.562640e-01
#>  [6,] 6.148986e-02 5.761387e-01 2.862624e-01
#>  [7,] 1.171032e-16 1.000000e+00 2.622462e-17
#>  [8,] 5.266130e-01 1.077283e-01 1.604470e-01
#>  [9,] 5.588728e-17 9.859033e-18 1.000000e+00
#> [10,] 1.075593e-01 5.646902e-01 1.886078e-01
bary(ems_pca, x)
#>                [,1]          [,2]         [,3]
#>  [1,] -0.0008319302  6.320082e-01 3.688237e-01
#>  [2,]  0.4778597220  1.949952e-01 3.271451e-01
#>  [3,]  1.0000000000  0.000000e+00 0.000000e+00
#>  [4,]  0.3641838837  3.908607e-01 2.449555e-01
#>  [5,]  0.5035764061  1.871646e-01 3.092590e-01
#>  [6,]  0.0541159317  6.336887e-01 3.121954e-01
#>  [7,]  0.0000000000  1.000000e+00 2.258079e-17
#>  [8,]  0.6385098515  1.564022e-01 2.050880e-01
#>  [9,]  0.0000000000 -3.862056e-17 1.000000e+00
#> [10,]  0.1632957053  6.228286e-01 2.138757e-01