Iterated Constrained endmembers

This method iteratively computes endmembers and abundances using quadratic programming until it converges to a solution.

ice(data, p, mu = 1e-05, t = 0.9999)

Arguments

data

Data matrix. Samples in rows frequencies in columns.

p

Number of endmembers.

mu

Regularization parameter from 0 to 1 that penaltizes the model for large simplex volume. The smaller the value the bigger the simplex

t

Tolerance ratio from 0 to 1 that affects number of iterations. The higher the value the more iterations.

Value

Structure with endmembers and abundances such that \(abundances * endmembers = data\_\)

where data\_ is an approximation of data. Endmembers are a matrix with samples in rows frequencies in columns.

References

M. Berman, H. Kiiveri, R. Lagerstrom, A. Ernst, R. Dunne, and J. F. Huntington, "Ice: A statistical approach to identifying endmembers in hyperspectral images: Learning from Earth's shapes and colors," IEEE Trans. Geosci. Remote Sens., vol. 42, no. 10, pp. 2085-2095, Oct. 2004.

E. M. Sigurdsson, A. Plaza and J. A. Benediktsson, "GPU Implementation of Iterative-Constrained Endmember Extraction from Remotely Sensed Hyperspectral Images," in IEEE Journal of Selected Topics in Applied Earth Observations and Remote Sensing, vol. 8, no. 6, pp. 2939-2949, June 2015. doi: 10.1109/JSTARS.2015.2441699